+++ packages/web3-providers/src/X2Y2/index.ts
@@ -1,4 +1,5 @@
 import urlcat from 'urlcat'
+import { first, last } from 'lodash-es'
 import {
     OrderSide,
     HubOptions,
@@ -16,8 +17,7 @@ import { ChainId, createERC20Token, createNativeToken, isZeroAddress, SchemaType
 import type { NonFungibleTokenAPI } from '../types/index.js'
 import { X2Y2_API_URL, X2Y2_PAGE_SIZE } from './constants.js'
 import type { Contract, Event, Order } from './types.js'
-import { first, last } from 'lodash-es'
-import { resolveActivityType } from '../helpers.js'
+import { resolveActivityType } from '../helpers/resolveActivityType.js'
 
 async function fetchFromX2Y2<T>(pathname: string) {
     const response = await globalThis.fetch(urlcat(X2Y2_API_URL, pathname))
