+++ packages/web3-providers/src/Zerion/index.ts
@@ -20,7 +20,6 @@ import { ChainId, createNativeToken, GasOption, SchemaType, isValidChainId } fro
 import type { ZerionNonFungibleTokenItem, ZerionNonFungibleCollection, ZerionCoin } from './types.js'
 import { formatAsset, formatTransactions } from './helpers.js'
 import type { FungibleTokenAPI, GasOptionAPI, HistoryAPI, NonFungibleTokenAPI, TrendingAPI } from '../types/index.js'
-import { getAllEVMNativeAssets, getAssetFullName } from '../helpers.js'
 import {
     getAssetsList,
     getCoinsByKeyword,
@@ -31,6 +30,8 @@ import {
     getTransactionList,
     zerionChainIdResolver,
 } from './base-api.js'
+import { getNativeAssets } from '../helpers/getNativeAssets.js'
+import { getAssetFullName } from '../helpers/getAssetFullName.js'
 
 const ZERION_NFT_DETAIL_URL = 'https://app.zerion.io/nfts/'
 const filterAssetType = ['compound', 'trash', 'uniswap', 'uniswap-v2', 'nft']
@@ -59,7 +60,7 @@ export class ZerionAPI
         return createPageable(
             unionWith(
                 assets,
-                getAllEVMNativeAssets(),
+                getNativeAssets(),
                 (a, z) => isSameAddress(a.address, z.address) && a.chainId === z.chainId,
             ),
             createIndicator(options?.indicator),
