+++ packages/web3-providers/src/helpers/fetch.ts
@@ -1,6 +1,6 @@
 const { fetch: originalFetch } = globalThis
 
-export type Fetcher = (input: RequestInfo | URL, init?: RequestInit, next?: Fetcher) => Promise<Response>
+export type Fetcher<T = Response> = (input: RequestInfo | URL, init?: RequestInit, next?: Fetcher) => Promise<T>
 
 export function fetch(input: RequestInfo | URL, init?: RequestInit, fetchers: Fetcher[] = []) {
     if (!fetchers.length) throw new Error('No fetcher found.')
