+++ packages/web3-providers/src/helpers/fetchR2D2.ts
@@ -1,3 +1,4 @@
+/* cspell:disable */
 import { attemptUntil, fetchImageViaDOM } from '@masknet/web3-shared-base'
 import { fetchBlob } from './fetchBlob.js'
