+++ packages/web3-providers/src/helpers/getNativeAssets.ts
@@ -1,0 +1,9 @@
+import type { FungibleAsset } from '@masknet/web3-shared-base'
+import { ChainId, createNativeToken, NETWORK_DESCRIPTORS, SchemaType } from '@masknet/web3-shared-evm'
+
+export function getNativeAssets(): Array<FungibleAsset<ChainId, SchemaType>> {
+    return NETWORK_DESCRIPTORS.filter((x) => x.isMainnet).map((x) => ({
+        ...createNativeToken(x.chainId),
+        balance: '0',
+    }))
+}
